# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LocalAI Dev Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deployments:
    get:
      operationId: listDeployments
      tags: [deployments]
      summary: List deployments
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
    post:
      operationId: createDeployment
      tags: [deployments]
      summary: Trigger a new deployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentInput"
      responses:
        "201":
          description: Created deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
...
  /containers:
    get:
      operationId: listContainers
      tags: [containers]
      summary: List running containers
      responses:
        "200":
          description: List of containers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Container"
    post:
      operationId: createContainer
      tags: [containers]
      summary: Create a new container
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContainerInput"
      responses:
        "201":
          description: Created container
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Container"
...
components:
  schemas:
...
    Deployment:
      type: object
      required: [id, workspaceId, status, environment, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        workspaceName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, building, deploying, success, failed, rolled_back]
        environment:
          type: string
          enum: [development, staging, production]
        url:
          type: ["string", "null"]
        commitSha:
          type: ["string", "null"]
        commitMessage:
          type: ["string", "null"]
        buildDuration:
          type: ["integer", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
        finishedAt:
          type: ["string", "null"]
    DeploymentInput:
      type: object
      required: [workspaceId, environment]
      properties:
        workspaceId:
          type: integer
        environment:
          type: string
          enum: [development, staging, production]
        commitSha:
          type: string
        commitMessage:
          type: string

    # ── Containers ────────────────────────────────────
    Container:
      type: object
      required: [id, workspaceId, name, image, status, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        name:
          type: string
        image:
          type: string
        status:
          type: string
          enum: [running, stopped, paused, exited, creating]
        port:
          type: ["integer", "null"]
        cpuPercent:
          type: ["number", "null"]
        memoryMb:
          type: ["integer", "null"]
        createdAt:
          type: string
        startedAt:
          type: ["string", "null"]
    ContainerInput:
      type: object
      required: [workspaceId, name, image]
      properties:
        workspaceId:
          type: integer
        name:
          type: string
        image:
          type: string
        port:
          type: integer
        envVars:
          type: object
          additionalProperties:
            type: string
...